Separate everything into their own files
Ran into circular dependency issues. Just figured I'd try to clean it up a little.
