+++ canvas/command/__init__.py
@@ -1,4 +1,7 @@
 from __future__ import annotations
 
 # flake8: noqa
-from .command import *
+from .parse import *
+from .factory import *
+from .base import *
+from .init import *
